Several skills and VampireModule system
Majority of skills planned have been added. Note: Testing still needs to be done, and some assets need to be made (sfx, videos, icons).

VampireModules are attached to the same gameObject that the Creature and Vampire script share. Based on the skills that a Vampire Creature possesses, these modules are attached to them. These modules allow for easier handling of skill functionality, especially when needing to use coroutines or access Vampire script members